mete el printf del campus

+++ src/libft/Printf/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: pjimenez <pjimenez@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 17:10:04 by pjimenez          #+#    #+#             */
-/*   Updated: 2023/06/20 17:33:00 by pjimenez         ###   ########.fr       */
+/*   Updated: 2024/01/30 11:15:14 by pjimenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
