+++ src/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: pjimenez <pjimenez@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 17:15:15 by pjimenez          #+#    #+#             */
-/*   Updated: 2024/01/19 16:35:17 by pjimenez         ###   ########.fr       */
+/*   Updated: 2024/01/30 11:17:59 by pjimenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int					ft_isdigit(int c);
 int					ft_isalnum(int c);
 int					ft_isascii(int c);
 int					ft_isprint(int c);
-size_t				ft_strlen(const char *str);
+int					ft_strlen(const char *str);
 void				*ft_memcpy(void *dst, const void *src, size_t n);
 size_t				ft_strlcpy(char *dest, const char *src, size_t size);
 int					ft_toupper(int c);
@@ -49,7 +49,7 @@ void				*ft_calloc(size_t count, size_t size);
 char				*ft_strdup(const char *s1);
 char				*ft_strjoin(char const *s1, char const *s2);
 void				*ft_memset(void *b, int c, size_t len);
-char				*ft_substr(char const *s, unsigned int start, size_t len);
+char				*ft_substr(char const *s, int start, int len);
 char				*ft_strtrim(char const *s1, char const *set);
 char				**ft_split(char const *s, char c);
 char				*ft_itoa(int n);
